requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
asyncpg==0.29.0
attrs==23.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
Cython==3.0.8
fasteners==0.19
filelock==3.12.4
names==0.3.0
prettytable==3.9.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
pytest==7.4.4
pytest-asyncio==0.23.3
python-docx==1.1.0
requests==2.31.0
setuptools==68.2.2
soupsieve==2.5
SQLAlchemy==2.0.25
sqlparse==0.4.4
googletrans==3.1.0a0
ujson==5.9.0
virtualenv==20.24.5
python-dotenv==1.0.1
greenlet==3.0.3
alembic==1.13.2
